requirements.txt
requests==2.25.1
requests-oauthlib==1.3.0
Flask==2.0.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.3
urllib3==1.26.3
Pillow==8.1.2
configparser==5.0.2